Update Switch to Follow New Architecture
