+++ Assets/Prefabs/Switch.prefab
@@ -10,7 +10,8 @@ GameObject:
   m_Component:
   - component: {fileID: 7540661724995432875}
   - component: {fileID: 1434380467501324467}
-  - component: {fileID: 4048976043567289606}
+  - component: {fileID: -2479629572066503108}
+  - component: {fileID: 1044000980393730825}
   m_Layer: 0
   m_Name: Switch
   m_TagString: Untagged
@@ -54,7 +55,7 @@ BoxCollider:
   serializedVersion: 3
   m_Size: {x: 2, y: 1, z: 2}
   m_Center: {x: 0, y: 0, z: 0}
---- !u!114 &4048976043567289606
+--- !u!114 &-2479629572066503108
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -66,9 +67,29 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 74d58d7e791b4c718d38d26b7b344ca9, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  modelData:
-    id: Switch_01
-    isOn: 0
-    preventTurningOff: 1
-  animator: {fileID: 0}
+  id: 0
+  isOn: 0
+  preventTurningOff: 0
   animationParamName: IsOn
+--- !u!95 &1044000980393730825
+Animator:
+  serializedVersion: 7
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2632030700321105197}
+  m_Enabled: 1
+  m_Avatar: {fileID: 0}
+  m_Controller: {fileID: 0}
+  m_CullingMode: 0
+  m_UpdateMode: 0
+  m_ApplyRootMotion: 0
+  m_LinearVelocityBlending: 0
+  m_StabilizeFeet: 0
+  m_AnimatePhysics: 0
+  m_WarningMessage: 
+  m_HasTransformHierarchy: 1
+  m_AllowConstantClipSamplingOptimization: 1
+  m_KeepAnimatorStateOnDisable: 0
+  m_WriteDefaultValuesOnDisable: 0
