--- Assets/Scripts/Environment/IInteractable.cs
@@ -1,10 +1,0 @@
-using UnityEngine;
-
-namespace Environment
-{
-    public interface IInteractable
-    {
-        void Interact();
-        Transform GetTransform();
-    }
-}
