--- Assets/Scripts/Environment/IInteractable.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 8f836f0d53934207bdcb17cb3826f3dc
-timeCreated: 1770139047
