+++ Assets/Scripts/Environment/InteractableObjects.meta
