+++ Assets/Scripts/Environment/InteractableObjects/InteractableBase.cs
@@ -1,0 +1,11 @@
+using UnityEngine;
+
+public enum InteractableID
+{
+    //TODO: Add IDs
+}
+
+public abstract class InteractableBase: MonoBehaviour
+{
+    [SerializeField] protected InteractableID id;
+}
