+++ Assets/Scripts/Environment/InteractableObjects/InteractableBase.cs.meta
