+++ Assets/Scripts/Environment/InteractableObjects/InteractableZone.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 6dce47a7f79a3213d9d0cb039b13b587
