--- Assets/Scripts/Environment/Switch/SwitchController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 9a4ee170acb640ecbe89360b1b16176f
