--- Assets/Scripts/Environment/Switch/SwitchModel.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: e0a52b1b5fd24856a403b4ebdb4c9ced
